功能：暂停Customer功能开发；完成remove对Foreign Key的判断

+++ src/main/java/team/wy/orchestra/dao/PlaysInstrDao.java
@@ -53,7 +53,7 @@ public class PlaysInstrDao {
         return playsInstrs;
     }
 
-    public PlaysInstr getById(int id) throws SQLException {
+    public PlaysInstr getById(long id) throws SQLException {
         Connection conn = DBHelper.getConnection();
         String sql = "select * from plays where id = ?";
         PlaysInstr playsInstr = runner.query(conn, sql, new BeanHandler<PlaysInstr>(PlaysInstr.class), id);
